[SQUASH] Refactor report gen
