+++ synapse/src/main/java/synapse/converter/report/Project.java
@@ -1,0 +1,30 @@
+/*
+ *  Copyright (c) 2025, WSO2 LLC. (http://www.wso2.com).
+ *
+ *  WSO2 LLC. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+
+package synapse.converter.report;
+
+import java.util.Arrays;
+
+public record Project(double overallConfidence, Mediator[] mediators) {
+
+    double timeEstimate(double baseTime) {
+        return Arrays.stream(mediators)
+                .mapToDouble(m -> m.timeEstimate(baseTime))
+                .sum();
+    }
+}
